requirements.txt
django==4.2.13
djangorestframework==3.15.1
psycopg2-binary==2.9.9
pyjwt==2.8.0
python-dotenv==1.0.1
pytest==8.2.2
pytest-django==4.8.0
pytest-env==1.1.3
allure-pytest==2.13.5